+++ tests/MetaBoxes/MetaBoxTwigViewTest.php
@@ -29,13 +29,13 @@ class MetaBoxTwigViewTest extends \WP_UnitTestCase
         ob_start();
         $stub->setTemplate('title');
         $stub->render($metaBox);
-        $this->assertEquals('My First Car', ob_get_contents());
+        $this->assertSame('My First Car', ob_get_contents());
         ob_end_clean();
 
         ob_start();
         $stub->setTemplate('min');
         $stub->render($metaBox);
-        $this->assertEquals('1', ob_get_contents());
+        $this->assertSame('1', ob_get_contents());
         ob_end_clean();
     }
 
@@ -46,12 +46,12 @@ class MetaBoxTwigViewTest extends \WP_UnitTestCase
          */
         $stub = new MetaBoxTwigView();
 
-        $this->assertEquals(null, $stub->getTemplate());
+        $this->assertSame('', $stub->getTemplate());
 
         $value = 'test-template.twig.html';
 
-        $this->assertEquals($stub, $stub->setTemplate($value));
-        $this->assertEquals($value, $stub->getTemplate());
+        $this->assertSame($stub, $stub->setTemplate($value));
+        $this->assertSame($value, $stub->getTemplate());
     }
 
     public function testGetterAndSetterContext()
@@ -61,14 +61,14 @@ class MetaBoxTwigViewTest extends \WP_UnitTestCase
          */
         $stub = new MetaBoxTwigView();
 
-        $this->assertEquals(array(), $stub->getContext());
+        $this->assertSame(array(), $stub->getContext());
 
         $value = array(
             'some_key' => 'some value',
         );
 
-        $this->assertEquals($stub, $stub->setContext($value));
-        $this->assertEquals($value, $stub->getContext());
+        $this->assertSame($stub, $stub->setContext($value));
+        $this->assertSame($value, $stub->getContext());
     }
 
     public function testGetterAndSetterTwigEnvironment()
@@ -78,7 +78,7 @@ class MetaBoxTwigViewTest extends \WP_UnitTestCase
          */
         $stub = new MetaBoxTwigView();
 
-        $this->assertEquals(null, $stub->getTwigEnvironment());
+        $this->assertSame(null, $stub->getTwigEnvironment());
 
         $value = new \Twig_Environment(
             new \Twig_Loader_Array(array(
@@ -86,7 +86,7 @@ class MetaBoxTwigViewTest extends \WP_UnitTestCase
             ))
         );
 
-        $this->assertEquals($stub, $stub->setTwigEnvironment(($value)));
-        $this->assertEquals($value, $stub->getTwigEnvironment());
+        $this->assertSame($stub, $stub->setTwigEnvironment(($value)));
+        $this->assertSame($value, $stub->getTwigEnvironment());
     }
 }