+++ tests/Transients/Special/BoolTransientTest.php
@@ -51,7 +51,7 @@ class BoolTransientTest extends \WP_UnitTestCase
             }
         } else {
             $this->stub->flush();
-            $this->assertEquals($expected, $this->stub->get());
+            $this->assertSame($expected, $this->stub->get());
         }
     }
 
@@ -71,9 +71,9 @@ class BoolTransientTest extends \WP_UnitTestCase
         $this->stub->set($value);
 
         if (class_exists($expected)) {
-            $this->assertEquals($value, $this->stub->get());
+            $this->assertSame($value, $this->stub->get());
         } else {
-            $this->assertEquals($expected, $this->stub->get());
+            $this->assertSame($expected, $this->stub->get());
         }
     }
 
@@ -84,6 +84,6 @@ class BoolTransientTest extends \WP_UnitTestCase
 
     public function testDefaultValue()
     {
-        $this->assertEquals(true, $this->stub->get());
+        $this->assertSame(true, $this->stub->get());
     }
 }