+++ tests/Transients/Special/DateTimeTransientTest.php
@@ -48,7 +48,9 @@ class DateTimeTransientTest extends \WP_UnitTestCase
                 try {
                     $this->stub->flush();
                 } catch (\Exception $exception) {
-                    $this->assertTrue(is_a($exception, $expected));
+                    $this->assertInstanceOf($expected, $exception);
+                } finally {
+                    $this->assertInstanceOf($expected, $exception);
                 }
             }
         }
@@ -62,6 +64,6 @@ class DateTimeTransientTest extends \WP_UnitTestCase
     public function testNull()
     {
         $this->stub->set(null);
-        $this->assertEquals('', $this->stub->get());
+        $this->assertNull($this->stub->get());
     }
 }